+++ src/cmd.rs
@@ -43,7 +43,8 @@ where
 
     match output {
         Ok(output) if output.status.success() => Ok(String::from_utf8(output.stdout)?),
-        _ => Err(eyre!("command failed: {:?}", args.into_iter().join(" "))),
+        Ok(output) => Err(eyre!("command failed: {:?}, exit_status_code: {:?}", args.into_iter().join(" "), output.status)),
+        Err(err) => Err(eyre!("command failed: {:?}, error: {:?}", args.into_iter().join(" "), err)),
     }
 }
 
@@ -77,7 +78,8 @@ where
 
     match status {
         Ok(status) if status.success() => Ok(()),
-        _ => Err(eyre!("command failed: {:?}", args.into_iter().join(" "))),
+        Ok(status) => Err(eyre!("command failed: {:?}, exit_status_code: {:?}", args.into_iter().join(" "), status)),
+        Err(err) => Err(eyre!("command failed: {:?}, error: {:?}", args.into_iter().join(" "), err)),
     }
 }